+++ src/crawlee/errors.py
@@ -15,6 +15,7 @@ __all__ = [
     'RequestHandlerError',
     'SessionError',
     'UserDefinedErrorHandlerError',
+    'ServiceConflictError',
 ]
 
 TCrawlingContext = TypeVar('TCrawlingContext', bound=BasicCrawlingContext, default=BasicCrawlingContext)
@@ -71,3 +72,12 @@ class ContextPipelineFinalizationError(Exception):
 
 class ContextPipelineInterruptedError(Exception):
     """May be thrown in the initialization phase of a middleware to signal that the request should not be processed."""
+
+
+class ServiceConflictError(RuntimeError):
+    """Thrown when a service container is getting reconfigured."""
+
+    def __init__(self, service_name: str, new_value: object, old_value: object) -> None:
+        super().__init__(
+            f"Service '{service_name}' was already set (existing value is '{old_value}', new value is '{new_value}')."
+        )
